Move from i32 to i64

+++ pkgs/backend/src/entities/wms/items.rs
@@ -6,12 +6,12 @@ use sea_orm::entity::prelude::*;
 #[sea_orm(table_name = "wms_items")]
 pub struct Model {
     #[sea_orm(primary_key)]
-    pub id: i32,
-    pub tenant_id: i32,
-    pub stock_id: i32,
-    pub lot_id: i32,
-    pub shelf_id: Option<i32>,
-    pub order_id: Option<i32>,
+    pub id: i64,
+    pub tenant_id: i64,
+    pub stock_id: i64,
+    pub lot_id: i64,
+    pub shelf_id: Option<i64>,
+    pub order_id: Option<i64>,
     pub expired_at: Option<DateTimeUtc>,
     pub created_at: DateTimeUtc,
     pub updated_at: DateTimeUtc,
