+++ pkgs/backend/src/entities/wms/picking_plans_in_zones.rs
@@ -6,9 +6,9 @@ use sea_orm::entity::prelude::*;
 #[sea_orm(table_name = "wms_picking_plans_in_zones")]
 pub struct Model {
     #[sea_orm(primary_key)]
-    pub id: i32,
-    pub tenant_id: i32,
-    pub plan_id: i32,
+    pub id: i64,
+    pub tenant_id: i64,
+    pub plan_id: i64,
     pub created_at: DateTimeUtc,
     pub updated_at: DateTimeUtc,
 }
